Revert "evergreen: Prep the updater for Cobalt (#7811)"
This reverts commit 19791ffc5227a1829c5229dc5834f93c1cdf814b.
