+++ cobalt/updater/BUILD.gn
@@ -32,6 +32,7 @@ static_library("updater") {
 
   deps = [
     "//base",
+    "//cobalt/browser:switches",
     "//components/crx_file",
     "//components/prefs",
     "//components/update_client",
