+++ cobalt/updater/configurator.cc
@@ -18,8 +18,8 @@
 #include <set>
 #include <utility>
 #include "base/command_line.h"
-#include "base/time/time.h"
 #include "base/version.h"
+#include "cobalt/browser/switches.h"
 #include "cobalt/updater/network_fetcher.h"
 #include "cobalt/updater/patcher.h"
 #include "cobalt/updater/prefs.h"
@@ -46,14 +46,6 @@ const char kUpdaterJSONDefaultUrlQA[] =
     "https://omaha-qa.sandbox.google.com/service/update2/json";
 const char kUpdaterJSONDefaultUrl[] =
     "https://tools.google.com/service/update2/json";
-
-// Whether to request, download, and install uncompressed (rather than
-// compressed) Evergreen binaries.
-const char kUseUncompressedUpdates[] = "use_uncompressed_updates";
-
-// Uses the QA update server to test the changes to the configuration of the
-// PROD update server.
-const char kUseQAUpdateServer[] = "use_qa_update_server";
 
 std::string GetDeviceProperty(SbSystemPropertyId id) {
   char value[kSystemPropertyMaxLength];
@@ -94,7 +86,7 @@ Configurator::Configurator(
     SetChannel(persisted_channel);
   }
   if (base::CommandLine::ForCurrentProcess()->HasSwitch(
-          kUseUncompressedUpdates)) {
+          switches::kUseUncompressedUpdates)) {
     use_compressed_updates_.store(false);
   } else {
     use_compressed_updates_.store(true);
@@ -121,13 +113,13 @@ base::TimeDelta Configurator::UpdateDelay() const {
 }
 
 std::vector<GURL> Configurator::UpdateUrl() const {
-#if !BUILDFLAG(COBALT_IS_RELEASE_BUILD)
-  base::AutoLock auto_lock(const_cast<base::Lock&>(update_server_url_lock_));
+#if !defined(COBALT_BUILD_TYPE_GOLD)
   if (allow_self_signed_packages_ && !update_server_url_.empty()) {
     return std::vector<GURL>{GURL(update_server_url_)};
   }
-#endif  // !BUILDFLAG(COBALT_IS_RELEASE_BUILD)
-  if (base::CommandLine::ForCurrentProcess()->HasSwitch(kUseQAUpdateServer)) {
+#endif  // !defined(COBALT_BUILD_TYPE_GOLD)
+  if (base::CommandLine::ForCurrentProcess()->HasSwitch(
+          switches::kUseQAUpdateServer)) {
     return std::vector<GURL>{GURL(kUpdaterJSONDefaultUrlQA)};
   } else {
     return std::vector<GURL>{GURL(kUpdaterJSONDefaultUrl)};
@@ -312,11 +304,11 @@ std::string Configurator::GetAppGuidHelper(const std::string& updater_channel,
             << "combination is undefined with the new Omaha configs.";
 
   // All undefined channel requests go to prod configs except for static
-  // channel requests for C24 and older.
+  // channel requestsf for C24 and older.
   // TODO(b/449024263): Replace regex matchers with substring_set_matcher or re2
   if (!std::regex_match(updater_channel, std::regex("2[0-4]lts\\d+")) &&
       sb_version >= 14 && sb_version <= 16) {
-    it = kChannelAndSbVersionToOmahaIdMap.find(
+    const auto it = kChannelAndSbVersionToOmahaIdMap.find(
         "prod" + std::to_string(sb_version));
     if (it != kChannelAndSbVersionToOmahaIdMap.end()) {
       return it->second;
@@ -327,7 +319,6 @@ std::string Configurator::GetAppGuidHelper(const std::string& updater_channel,
 }
 
 std::string Configurator::GetAppGuid() const {
-  base::AutoLock auto_lock(const_cast<base::Lock&>(updater_channel_lock_));
   const std::string version(COBALT_VERSION);
   return GetAppGuidHelper(updater_channel_, version, SB_API_VERSION);
 }
