+++ cobalt/updater/network_fetcher.cc
@@ -189,10 +189,8 @@ void NetworkFetcher::DownloadToFile(
       url_loader_factory_.get(),
       base::BindOnce(
           [](const network::SimpleURLLoader* simple_url_loader,
-             DownloadToFileCompleteCallback download_to_file_complete_callback,
-             // A base::FilePath parameter is needed to match the callback
-             // definition.
-             base::FilePath file_path) {
+             DownloadToFileCompleteCallback
+                 download_to_file_complete_callback) {
             std::move(download_to_file_complete_callback)
                 .Run(simple_url_loader->NetError(),
                      simple_url_loader->GetContentSize());
