+++ components/update_client/network.h
@@ -76,10 +76,6 @@ class NetworkFetcher {
       ProgressCallback progress_callback,
       DownloadToFileCompleteCallback download_to_file_complete_callback) = 0;
 
-#if BUILDFLAG(IS_STARBOARD)
-  virtual void Cancel() = 0;
-#endif
-
  protected:
   NetworkFetcher() = default;
 };
