+++ components/update_client/url_fetcher_downloader.cc
@@ -28,7 +28,7 @@
 
 namespace {
 
-#if BUILDFLAG(IS_STARBOARD) && !defined(IN_MEMORY_UPDATES)
+#if BUILDFLAG(IS_STARBOARD)
 // Can't simply use base::DeletePathRecursively() because the empty dirs
 // need to be preserved.
 void CleanupDirectory(base::FilePath& dir) {
@@ -126,9 +126,6 @@ void UrlFetcherDownloader::SelectSlot(const GURL& url) {
     ReportDownloadFailure(url, CrxDownloaderError::SLOT_UNAVAILABLE);
     return;
   }
-  config_->SetUpdaterStatus(std::string(
-      updater_status_string_map.find(UpdaterStatus::kSlotLocked)->second));
-
   // Use 15 sec delay to allow for other updaters/loaders to settle down.
   base::SequencedTaskRunner::GetCurrentDefault()->PostDelayedTask(
       FROM_HERE,
@@ -189,9 +186,11 @@ void UrlFetcherDownloader::DoCancelDownload() {
   LOG(INFO) << "UrlFetcherDownloader::DoCancelDownload";
   DCHECK_CALLED_ON_VALID_SEQUENCE(sequence_checker_);
   is_cancelled_ = true;
-  if (network_fetcher_.get()) {
-    network_fetcher_->Cancel();
-  }
+  // TODO(b/431862767): enable this in a follow-up PR with the Cobalt network fetcher implementation
+  NOTIMPLEMENTED();
+  // if (network_fetcher_.get()) {
+  //   network_fetcher_->Cancel();
+  // }
 }
 #endif
 
@@ -269,7 +268,6 @@ void UrlFetcherDownloader::StartURLFetch(const GURL& url) {
       base::BindOnce(&UrlFetcherDownloader::OnNetworkFetcherComplete, this));
 #else
   file_path_ = download_dir_.AppendASCII(url.ExtractFileName());
-  LOG(INFO) << "UrlFetcherDownloader::StartURLFetch, file_path_ =" << file_path_.value().c_str();
   network_fetcher_->DownloadToFile(
       url, file_path_,
       base::BindOnce(&UrlFetcherDownloader::OnResponseStarted, this),
